Website_Layout/WINE/pyproject.toml
[tool.poetry]
name = "wine"
version = "0.1.0"
description = ""
authors = ["dvmn-tasks <devman-tasks@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Jinja2 = "^3.1.2"
pandas = "^1.5.2"
openpyxl = "^3.0.10"
pep8 = "^1.7.1"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
